Fix macOS compilation errors

+++ reg-test/reg_test_voxelCentricToNodeCentric.cpp
@@ -160,7 +160,7 @@ public:
                     reg_mat44_mul(&transformation, nodeCoord, voxelCoord);
                     // Linear interpolation
                     DataType basisX[2], basisY[2], basisZ[2];
-                    const int pre[3] = { reg_floor(voxelCoord[0]), reg_floor(voxelCoord[1]), reg_floor(voxelCoord[2]) };
+                    const int pre[3] = { (int)reg_floor(voxelCoord[0]), (int)reg_floor(voxelCoord[1]), (int)reg_floor(voxelCoord[2]) };
                     basisX[1] = voxelCoord[0] - static_cast<DataType>(pre[0]);
                     basisX[0] = static_cast<DataType>(1) - basisX[1];
                     basisY[1] = voxelCoord[1] - static_cast<DataType>(pre[1]);
